my version upload
This is a super stripped up version of vrp2 for vrp3. I took removed practically everything except the tunnel and proxy, and then created this version. It's not at all complete, no where near, but it has a character ui, and working vrp2 self remote and events.

There is no GUI yet. I think it may be a separate resource, or brought in at a later date.